spark/test: move DynamicParameter to a shared test-commons module
Extract the DynamicParameter test helper into a new test-commons module exposed
as test fixtures, and rename the EMR property prefix from 'openlineage.test' to
'openlineage.tests.emr'.

BREAKING: external users setting EMR integration-test parameters via the
'openlineage.test' system-property prefix must switch to 'openlineage.tests.emr'.
