+++ integration/spark/app/src/test/java/io/openlineage/spark/agent/EmrIntegrationTest.java
@@ -63,8 +63,8 @@ import org.junit.jupiter.api.Test;
  * <h2>Configuration</h2>
  *
  * <p>All infrastructure details and configuration parameters should be set using system properties.
- * For example: {@code -Dopenlineage.test.bucketName=my-bucket-name}. Most parameters have defaults.
- * For a full list of configurable parameters, see {@link EmrDynamicParameter}.
+ * For example: {@code -Dopenlineage.tests.emr.bucketName=my-bucket-name}. Most parameters have
+ * defaults. For a full list of configurable parameters, see {@link EmrDynamicParameter}.
  *
  * <h3>Note on JUnit and Gradle</h3>
  *
